fixed search bug, fixed guildkey bug, refactor sanitization component function
