+++ src/functions/index.js
@@ -1,1 +1,2 @@
 export { apiCall } from "./apiCall"
+export { sanitizeHtmlString } from "./sanitizeHtmlString"
